Make table metadata options class visible (JAVA-946).
