+++ changelog/README.md
@@ -9,6 +9,7 @@
 - [bug] Fix Configuration builder to allow disabled metrics (JAVA-900)
 - [new feature] Prepare API for async query trace (JAVA-902)
 - [new feature] Add BoundStatement#unset (JAVA-930)
+- [bug] Make table metadata options class visible (JAVA-946)
 
 
 ### 3.0.0-alpha3
